+++ src/Ampersand/Prototype/GenFrontend.hs
@@ -184,13 +184,13 @@ buildInterface :: (HasDirPrototype env) => FSpec -> [Interface] -> Interface -> 
 buildInterface fSpec allIfcs ifc = do
   obj <- buildObject (BxExpr $ ifcObj ifc)
   return 
-    FEInterface { ifcName = escapeIdentifier $ name ifc
-                , ifcLabel = name ifc
-                , _ifcExp = objExp obj
+    FEInterface { ifcName    = escapeIdentifier (name ifc)
+                , ifcLabel   = name ifc
+                , _ifcExp    = objExp obj
                 , _ifcSource = objSource obj
                 , _ifcTarget = objTarget obj
-                , _ifcRoles = ifcRoles ifc
-                , _ifcObj = obj
+                , _ifcRoles  = ifcRoles ifc
+                , _ifcObj    = obj
                 }
     -- NOTE: due to Amperand's interface data structure, expression, source, and target are taken from the root object. 
     --       (name comes from interface, but is equal to object name)